Update patch set 1
I would prefer that you didn't submit this.

This is a good addition. There are a few minor style/doc issues that are
easy to fix.

However, I agree with ralf that duplicating the build template is a
problem.

At a minimum, the same template (build.template) should be able to
generate both build.xml files.

A better solution would be a single build.xml in the main project able
to build the test project by reusing the targets of the various build
steps.

Patch-set: 1