Update patch set 1
Thanks for the comments. Working on the following fixes:

1. Fix style issues.
2. A single build.template to generate main/build.xml and
tests/build.xml

Lets keep the projects separate (main/build.xml & tests/build.xml) for
now, instead of a single build.xml file for both main and tests. We can
discuss if a single build.xml for both main/tests might be a better
option and make a decision based on that later.

On 2008/11/14 00:19:03, "Xavier Ducrohet" <xav wrote:
> I would prefer that you didn't submit this.
> 
> This is a good addition. There are a few minor style/doc issues that
are
> easy to fix.
> 
> However, I agree with ralf that duplicating the build template is a
> problem.
> 
> At a minimum, the same template (build.template) should be able to
> generate both build.xml files.
> 
> A better solution would be a single build.xml in the main project able
> to build the test project by reusing the targets of the various build
> steps.
> 
> 

Patch-set: 1