Update patch set 4
Just a note to apologise for all the mails regarding this change. (And this one too.)

I'm still learning git and repo. Which I know is no excuse for being a muppet. (I was just trying to remove a stray space from the end of one line of the patch).

Incidentally, if you have the power to abandon change 8874, please do. It seems that it can't be done through the web interface.

Thanks,

Adrian

Patch-set: 4