Update patch set 1
Patch Set 1: No score

Thanks for taking a look at the change. I added this feature, because we are using apkbuilder as part of an ANT build script to build and run our tests. The tests must be signed with the same custom debug key as the program. Why do we need to sign with a custom debug key, you might ask? Well, our application has to use a custom debug key so that it works with the Google Maps API key, and so that it works this way consistently for all of our developers. All of our developers use the custom debug keystore that we have checked into the repository.

Does that help clarify things? Thank you!

Patch-set: 1
Reviewer: Sangjin Yun <1002034@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0