Add simple hello-native example.
Copy hello-jni into hello-native and remove all the Java™ parts,
making it into a native “Hello World” sample application.

The process of writing an Android.mk for mksh involves letting
its Build.sh configure script compile and link a few “scan”
programmes, and for this we need appropriate definitions of
CC, CFLAGS, CPPFLAGS, LDFLAGS, and LIBS. We generate these
via system/core/sh/mkmf.sh which contains excerpts of the
build log of compiling this application. (It also is fully
functional as a console “Hello World”.)