Update patch set 11
Patch Set 11: No score

Actually, my recommendation would be to leave the change "as is" in the open-source tree, and as part of the merge to Google's internal master tree to skip the merge of that one file - the other files merge and build fine. This way the open-source branch will look as correct as practical for as long as practical.

Patch-set: 11