Update patch set 1
Patch Set 1:

Style looks fine. Merges and builds fine in the open-source tree, merges fine in Google's internal tree. Untested.

It'd be good if you get a chance to reformat the commit message. It's traditional in git to limit the first line to 65 characters followed by a blank line, and to use at most 70 characters on subsequent lines.

Patch-set: 1