Win SDK: Split the win-sdk makefile and script into 2 parts.
The main makefile+scripts to build the Windows SDK is still this
one here in development/build/tools. However it defers to a new
matching set of files in sdk/build to build and package
things that depend on the sdk.git or external/qemu.git.

This will make it easier for us to prepare SDKs based on
a tools_rN branch that isn't cut at the same time than the
platform branch.

This is a multi-part changeset. The other parts is in sdk.git.

Change-Id: I4e9a2d810cf29fae0097fbd92be0cef89c9b3505