Update patch set 5
Patch Set 5: Verified

For the record, the patch was verified with libgl1-mesa-dev installed on Ubuntu 10.10. I have also asked for the installation of the package on all build servers. Let's wait a bit to see if this happens quickly, and I'll +2 the change.

I.e. don't bother changing this yet. Thanks.

Patch-set: 5
Label: Verified=+1