emulator: opengl: ut_renderer: Events redirection.
This implements event redirection in the "ut_renderer" program.
What this does is redirect mouse clicks and a few key presses to
the emulator's console through localhost:5554.

Note that only a few key events are supported (see NativeWindowing.h)
since supporting everything is much more work, and not needed at this
point.

The sockets.[hc] and iolooper*[hc] source files were originally
written for the emulator, and are here relicensed under Apache 2.0
(I wrote 100% of them, so no issue here).

This is only for demo purposes. Only tested on Linux.

Change-Id: I2af068c2ef8c535b80fd4432c452c8ce0f1bfa27