Update patch set 3
Patch Set 3: Do not submit

Ok, I have uploaded the following patch to update the 32-bit prebuilt sysroot used for the SDK: https://review.source.android.com/22937

When it is applied, the patch still fails. It misses to list libGLcommon in its LOCAL_SHARED_LIBRARIES.

Also, please change the name to something else, "integration" is too generic, we need something that tells that this is a test program (e.g. test_GLES_translator or something).

Patch-set: 3