Monkey not handling SIGHUP signal
This fix resolves a problem that has always been present but not
always visible. A fix in bionic made the problem visible (see
ff3129b0fe52fabec273077871cb35515465d44b )
adb daemon on the device side terminates connection with the host
and sends SIGHUP signal to all its children involved in this
connection and also to the shell process being monkey's parent.
This fix is creating a new session for the Monkey process which
causes it to be detached from the controlling terminal and allows
it to run as a stand-alone application. The new session leader is
not affected by the SIGHUP signal sent by the parent process.

Change-Id: I2983a73b86a796a8cc8d29368e5c8b0a2e65c0ef