Update patch set 1
Patch Set 1:

I forgot to mention that I got it working for Swedish with several changes. Unfortunately the output isn't quite as good as it seems to be for English. Not too bad, but needs some polishing.

By far the biggest problem is that LatinIME tries to separate words when it shouldn't, ie. for any combined word. Ex: "tunnelbanan" becomes "tunnel banan" and "telefonnummer" becomes "telefon nummer". I understand that it depends on how different languages handles combined words though but it worked fine before ICS.

Patch-set: 1