Simple Tic Tac Toe app which demonstrates NFC sharing.
Android Beam allows simple peer-to-peer data exchange between two Android-powered devices.
The application that wants to beam data to another device must be in the foreground and the device receiving the data must not be locked.
When the beaming device comes in close enough contact with a receiving device, the beaming device displays the "Touch to Beam" UI. The user can then choose whether or not to beam the message to the receiving device.

It will demonstrate the following things -

1.How to Write to a NFC Tag.
2.How to Read from NFC Tag.
3.How to Exchange data between two NFC enabled devices having the same app installed.

Change-Id: I920fb25315c06cee6607b4710594048697ff3dd5
Signed-off-by: iBadrinath <urbaddy@gmail.com>