Update patch set 1
Patch Set 1:

There's no mistake here. The issues you mention are totally orthogonal.

One simple issue is that the Android driver distributed by Google should only reference products with the Google VID. Whether a device has this VID or the one of the OEM is a partnership/strategy decision and not a technical one.

The other issue is that the way Windows USB driver works, all possible devices need to be enumerated. Yet in this case the protocol behind is the same for all devices. That is if you modify your _own_ installation of the Android USB driver and add the missing line to your config file, you can load that and don't need an OEM Samsung driver. I do that all the time on my boxes.

As for your other points: the size of the driver depends on how they created it and what "features" they choose to add (e.g. I'd guess they need to support some of their own custom devices, which the Android USB driver doesn't). And the GSM vs CDMA thing is totally irrelevant too since the USB protocol doesn't change -- the documentation simply uses the links advertised by the OEMs.

I hope this explains the situation in more details. I don't make policies and this isn't the right medium to discuss them, so if you still feel like replying I'd invite you to find a relevant forum to do so.

Since there will be no non-Google VID added to this file, I'd recommend you abandon this patch. Thanks for submitting a CL to AOSP though; as an engineer I appreciate the effort.

Patch-set: 1