Update patch set 2
Patch Set 2:

What's with the duplicated PNG and SVG files? Do you intend the app to use the PNG or the SVG ones? If you want the SVG ones to be like "raw" assets and PNGs to be rendered versions of them, maybe they should be in different directories.

Patch-set: 2
Label: Code-Review=0