Update patch set 1
Patch Set 1: Code-Review+1

I'm approving this change as an optimization only.  It is worthwhile and will sometimes conceal the issue but it does not solve the root cause.

Michael will work on a fix for the real problem in the NDK wrappers.

Patch-set: 1
Reviewer: Jeff Brown <1003957@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1