Hold reference to old apps in AppListLoader#deliverResult()
This patch fixes a bug in the custom Loader implementation
provided in the ApiDemos. Currently, the AppListLoader's
deliverResult() method does not hold a reference to the
old apps as it should (this was most likely a typo).

These changes mirror the CursorLoader.java implementation,
so check the CursorLoader#deliverResult() method's source
code for more information.

Change-Id: Iab2736a5fd9aad83bf64d5eadea7336bb5a3bbcb
Signed-off-by: Alex Lockwood <alexjlockwood@gmail.com>