Update patch set 1
Patch Set 1:

I'm ok with this change, but it would really be better if there was a way to automate rebuilding the library from known sources / GIT commits. I'll try to come up with something.

For Chrome, we will likely want to rebuild the library from sources directly (because this allows other things, like instrumenting the code with ASan and other tools), but this should be nice for NDK developers.

Patch-set: 1
Label: Code-Review=0