ndk: Add missing declaration for mkdtemp.h.
The mkdtemp() function has always been implemented in the C library
(see [1] for proof) but was missing a declaration in <stdlib.h>, until
Android 4.0.1 [2]

This patch simply adds the missing declaration to former <stdlib.h>
versions. This makes it easier to build certain open-source programs
for Android, because auto-tools probing typically looks at the symbols
in libc.so to define HAVE_MKDTEMP, then the source fails to compile
because of the lack of declaration.

[1] https://android.googlesource.com/platform/bionic/+/android-1.6_r1/libc/stdio/mktemp.c
[2] https://android.googlesource.com/platform/bionic/+/ad1ff2fb268034074488f0b4b30b5311008b8617

Change-Id: I838b54bcb5cbfd2fed24cf7649897889ec7a5559