libportable: Add dependency on libdl.
The patch at [1] was reverted because ndk/sources/android/cpu-features
is now part of the platform build (as a dependency of libportable), and
now requires dlopen()/dlsym()/dlclose().

Because the platform build system doesn't support LOCAL_EXPORT_LDLIBS,
add the dependency to libportable's own module definition instead.
After this patch is submitted, it's possible to re-apply the original
patch and get a working build.

BUG=13679666

Change-Id: I54b4376a562242d21631eaf13090650decfe791d