Autodetect architectures based on register names.
This assumes that the full stack trace is pasted, starting
from the first line of debuggerd output. In practice, this isn't
as restrictive as it sounds because the web tool needs the build
ID to fetch symbols from the build server.

This change also fixes the broken regular expressions for debuggerd
register output.

- We no longer assume registers are 32 bits wide.
- We no longer assume there will be exactly 4 regs per line of output
- We no longer assume there will be exactly one space between the
  register name and its contents
- We no longer assume register names are lower case.

bug: 15084471

Change-Id: I8947dc122f692df93b8297eee730e914091ac9e5