Update patch set 3
Patch Set 3:

> Less than 2% of our users are on XP and Vista. Will this make Win7
 > the minimum requirement?

as spencer says, "probably not" --- it'll _probably_ still work on older releases. but, yes, it will mean Win7 is the minimum guaranteed to work.

we don't actually need to do this for the adb usb .dlls --- i did it by mistake because i'm using Win7 for testing. but we do need to bump to at least Vista for the higher-level code at some point. (apparently the libc++ folks are still arguing about whether to support Vista or 7 as their minimum.)

 > And when will this get into the release
 > branch?

that will take forever, as we discussed the other day. (that is: i have no plans to try to cherrypick anything into mnc-release.) i do wish Android Studio had different channels like Debian or ChromeOS; although i wouldn't want to inflict anything like AOSP ToT on _everyone_ it would be nice -- given that internally we basically only have Linux users -- to have some external users who aren't always a year behind. (even the impossible to find release candidates are pretty stale at this point!) but that's a separate discussion for another day :-)

Patch-set: 3