gdbrunner: fix background/stdin interaction.
gdbrunner starts the remote gdbserver process in the background which
disconnects it from stdin. However now that adb is reading stdin for
non-interactive shells, the background adb process gets hit with a
SIGTTIN and stops, which caused gdbclient.py to fail to receive
gdbserver output.

This fix adds an explicit background parameter to shell_popen(), and if
it's set disables reading from terminal stdin. This makes shell_popen()
a bit more flexible for general use and makes putting the process in
the background a little more foolproof.

Bug: http://b/25192785
Change-Id: Ieeae05262a4a16d37d50bcda41b67a71f37a7d0d