Update patch set 1
Patch Set 1:

OK I think I got it. You're right, it's the push that's failing, but the reason is that Powershell can't find my adb :\ I don't have adb on my PATH because I like to bounce between versions for testing, so I just use a local adb.exe in that directory.

It looks like both the subprocess module and the basic cmd.exe are happy to execute a program in the local directory, but Powershell doesn't allow it.

I'm not quite sure how to get around this. It's possible that we don't want to worry about supporting my workflow and we could just require adb to live on PATH, but it is really handy for me to be able to just change folders to get a different adb version. I don't see any options to tell Powershell to allow local execution though, and prefixing with .\ isn't an option because then it will break using the PATH adb. Do you know of any way to get Powershell to revert to cmd.exe behavior?

(btw I'm on vacation next week so I probably won't be checking back here, if you and enh@ come to a consensus feel free to move ahead without waiting for me :)

Patch-set: 1
Label: Code-Review=0