Style change; exit cleanly on success
- Quoted vars to prevent accidental word-splitting
- Some small style/best practice modifications
- Manually clean up and unset trap at end of script

The last one is the most significant as it may impact automation
involving the script (but really in a good way). This change will let a
calling script check the exit code to determine success or failure.

Due to the trap put in place midway through the script, a successful run of the
script was previously ended with a cleanup (good) and `exit 1` (not good).
A non-zero exit code indicates failure.

Emitting a non-zero (failure-indicating) exit code on success makes a script
difficult to work with from an automation perspective. In this case, it
leaves the caller to either A) parse make_key's output or B) verify
make_key's keys.

Change-Id: I5b7dcf892dc87f2b108e0469130ae6b103125add