Update patch set 1
Patch Set 1: Code-Review-1

Why only android-21? Shouldn't this go to the earlier versions of android too?

Also I would suggest doing this change on the top of the copy of the list from the previous version of android, to see which symbols were added and removed not just a list of new symbols.

Patch-set: 1
Reviewer: Dimitry Ivanov <1041673@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1