WidgetPreview: Add minSdkVersion to manifest
This app was not available before API 11, so its minSdkVersion
should be 11. Declaring minSdkVersion explicitly in the manifest is
important, because otherwise the AOSP build system will assign the
current platform SDK version to it, e.g. 23 for AOSP master or "N"
for nyc-dev. In the case of N, this will lead to (partly due to a
PackageParser bug [1]) the app not getting the required storage
permission automatically, because the system assumes it supports
runtime permissions. Consequently, the app will not work before the
user enables the storage permission for it in Settings.

[1] See https://android-review.googlesource.com/231353

Signed-off-by: Jinhui Li <jinhui.li@intel.com>
[Revised code and commit message]
Signed-off-by: Yu Ning <yu.ning@intel.com>

Change-Id: I4604af4fdb99eacbbadb922c6a0a89b4cec51dbb