KeyDerivationFunction: example about treating data encrypted via SHA1PRNG
The Crypto provider providing the SHA1PRNG algorithm for random number
generation was deprecated.

There was an usual anti-pattern of using that algorithm to derive keys.

This example provides a helper class and shows how to treat data that
was encrypted in the incorrect way and re-encrypt it in a proper way.

Bug: 27873296

Change-Id: I92d2fd4ebb07d5823de31f5a199e23b1dba4836e