Remove hard coding of ascii encoding when compiling SDK stubs
Hard coding the compiler to treat the Java files as if they
were ASCII is inconsistent with how Java is compiled everywhere
else in the Android build. This causes problems with the upgrade
to ICU 58 which adds copyright symbols in UTF-8 encoding in its
copyright comments.

This change has no discernable effect on the SDK generated from
the existing Java code.

Test: Built the SDK with and without and compared the two.
Change-Id: Ied5ba2fa98e1e0ca6c265ba353b4091892170541