Update patch set 1
Patch Set 1:

It lives in multiple places in the tree :(. There's a bug against build team to give us a central place to put it that can just mirror upstream and everything else can reference, but it's been languishing and I haven't had time to push it.

The impetus for this update is the change to vk_platform.h -- when building for some Android ABIs, you'll get warnings (or errors with -Werror) if you use -Wundef. If you want, I can just cherry-pick that fix manually until we get the process sorted out a bit better, but as much as possible I try to keep all of our copies unmodified from upstream.

Patch-set: 1
Label: Presubmit-Ready=0