Update patch set 3
Patch Set 3: Code-Review-2

I still disagree with this patch. Either you're trying to generate pointer events for a display that doesn't exist (which is either a script bug or a Monkey bug that should be fixed), or system_server is dead. Either way, the correct answer is to raise an error some how, not silently succeed.

Patch-set: 3
Reviewer: Michael Wright <1006497@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-2