Update patch set 2
Patch Set 2:

Hi Aurimas, sorry I didn't manage to *naturally* return to the CL that kicked off this discussion, so I cherry picked it again to get a sample. I've shared the process and the results. Perhaps you can take a look at the list of errors I get once I perform the indent and see what you think. (Link to the doc shared out-of-band). To me it looks like the only thing that it's still complaining about is:
1) other indents that are still wrong
2) intra-line whitespace
That puts the cost of being helpful somewhere in the neighborhood of 50 nits.

Patch-set: 2