WIP : Added header-abi-linker.
header-abi-linker links dumps produced by header-abi-dumper when it goes
over each source file in a vndk shared library. It simply looks at the
"mlinker_set_key" field of the protobuf message to do this.

Only abi information present in source files in "export_include_headers"
is dumped now. If a library does not export headers it's abi is not
dumped.

Test: header-abi-dumper -e tests -o example1.dump tests/example1.cpp --
-x c++ -std=c++11 -I . -I ~/android/aosp/external/clang/lib/Headers/

Change-Id: I692de0933349a6b80a6b1a3b605a377a25cd0b5a