vndk-def: Separate {sp-hal,sp-ndk}-vndk-stable
This commit separates the vndk-stable set, which was printed by vndk
subcommand, into sp-hal-vndk-stable and sp-ndk-vndk-stable.  This will
make it easier to diagnose the shared library usages.

Note: The sp-lib subcommand was printing separated vndk-stable sets even
before this commit.  This commit unifies the output format of vndk and
sp-lib subcommands.

Test: Run vndk subcommand over sailfish image.
Change-Id: I5b64d4359da8aa344dc87a5b3a8469c0030c115c