C/C++ adjustments and bug fixes.
Bug: 37751376

Test: dumped the abi of libminijail, removed the "struct" keyword from
the abi-dump and confirmed that this did not result in an abi breakage
getting reported.

Bug: 37749718

Test: manually invoked header-abi-diff with file containing a list of
symbols to omit abi checks on from libminijail after changing sizes and
alignments of those symbols from libminijail's abi dumps. Diffs were not
reported.

Bug: 37918686

Test: dumped the abi of libbcinfo, it contains records as well as
symbols which adhere to the regex in the symbol file.

Also make function signature diffing independent of C++ name mangling.

Change-Id: Id34209ef24d1202745ca69765a4e2a96d1a05c2a