NsdChat: Advertise unique service names
The existing code tries to register "NsdChat" on every node, then let
the framework resolve collisions itself.  It incorrectly assumed that
the local node would always be assigned "NsdChat" and that other nodes
on the LAN would be assigned different names.  This is a confusing way
to design a sample P2P app, and it didn't seem to work very well in
practice.

So instead, let's give each node a unique ID so there is no ambiguity
as to what the service name checks are doing.