Update patch set 9
Patch Set 9: Presubmit-Verified+1

=== Finished presubmit run: L47200000125910890 ===

Status: https://android-build.googleplex.com/builds/where-is-my-cl/android-review/552479?revision=9




TreeHugger finished with: 15 passed.







WARNING: The tip-of-tree(4489670) for git_master/gce_x86_phone-userdebug_fastbuild3c_linux may be broken.
The presubmit build passes with the last-known-good-build(4489602), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.









WARNING: The tip-of-tree(4489670) for git_master/marlin-userdebug_fastbuild_linux may be broken.
The presubmit build passes with the last-known-good-build(4489558), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.




WARNING: The tip-of-tree(4489670) for git_master/sailfish-userdebug_fastbuild3_linux may be broken.
The presubmit build passes with the last-known-good-build(4489602), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.






Impressive, young Jedi.

Have a question? Found a bug? http://go/treehugger-troubleshoot

Patch-set: 9
Tag: autogenerated:TreeHugger:Verify-Pass
Reviewer: Treehugger Robot <1062513@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Verified=+1