Update patch set 1
Patch Set 1:

I am strongly against this change, and I would -2 it if I could.

Please can you explain why this directory is so strongly locked down ? It contains a *lot* of things that are not the responsibility of any of the owners. If the folks advocating for a strong lockdown care about something very strongly, they should feel free to add an OWNERS file to a specific subdirectory of development. 

Exhibit A :I filed b/70201085 because people were facing unnecessary friction in getting changes to development/samples/ submitted.

Exhibit B : ndk/ is owned by danalbert@ / enh@. Why are they not in this list ?

Patch-set: 1
CC: Narayan Kamath <1014443@85c56323-6fa9-3386-8a01-6480fb634889>